Cleanup & hide functions will now ignore main Chrome window (if it has more than one tab)

+++ destroybots.py
@@ -2,31 +2,37 @@ import win32gui
 import win32process
 import win32api
 import time
+from utils.count_processes import count_processes
 from utils.logger import log
 
 start = time.perf_counter()
-
+counters = count_processes()
 
 def foreach_window(hwnd, lParam):
     global start
     title = win32gui.GetWindowText(hwnd).lower()
-    exe = ''
     if title.find("amazon") != -1:
         pid = win32process.GetWindowThreadProcessId(hwnd)
-        for p in pid:
-            try:
-                hndl = win32api.OpenProcess(0x0400 | 0x0010 | 0x0001, False, p)
-                exe = win32process.GetModuleFileNameEx(hndl, False)
-            except:
-                exe = ''
-            if exe == 'C:\\Program Files\\Google\\Chrome\\Application\\chrome.exe':
+        if len(pid) <= 2:
+            for p in pid:
                 try:
-                    win32api.TerminateProcess(hndl, -1)
-                    log.info(f'Terminated Chrome window {title}.')
-                    start = time.perf_counter()
-                except win32api.error:
-                    log.info(f'Unable to terminate process {title}.')
-                break
+                    count = counters[p]
+                except KeyError:
+                    count = 1
+                if count < 2:
+                    try:
+                        hndl = win32api.OpenProcess(0x0400 | 0x0010 | 0x0001, False, p)
+                        exe = win32process.GetModuleFileNameEx(hndl, False)
+                    except:
+                        exe = ''
+                    if exe == 'C:\\Program Files\\Google\\Chrome\\Application\\chrome.exe':
+                        try:
+                            win32api.TerminateProcess(hndl, -1)
+                            log.info(f'Terminated Chrome window {title}.')
+                            start = time.perf_counter()
+                        except win32api.error:
+                            log.info(f'Unable to terminate process {title}.')
+                        break
     return True
